+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
+- `moonstone/ContextualPopupDecorator` config option `openProp` to support mapping the value of `open` property to the chosen property of wrapped component
+
 ### Changed
 
 - `moonstone/Popup` to only call `onKeyDown` when there is a focused item in the `Popup`
@@ -28,7 +30,6 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/styles/text.less` mixin `.locale-japanese-line-break()` to apply the correct  Japanese language line-break rules for the following multi-line components: `moonstone/BodyText`, `moonstone/Dialog`, `moonstone/Notification`, `moonstone/Popup`, and `moonstone/Tooltip`
 - `moonstone/ContextualPopupDecorator` property `popupProps` to attach props to popup component
-- `moonstone/ContextualPopupDecorator` config option `openPropMap` to support mapping the value of `open` property to the chosen property of wrapped component
 - `moonstone/VideoPlayer` property `pauseAtEnd` to control forward/backward seeking
 - `moonstone/Panels/Header` prop `marqueeOn` to control marquee of header